+++ R/mentor.r
@@ -192,7 +192,7 @@ plot.diffnet_mentor <- function(
   x,
   y            = NULL,
   vertex.size  = "degree",
-  minmax.relative.size = getOption("diffnet.minmax.relative.size", c(0.025, 0.05)),
+  minmax.relative.size = getOption("diffnet.minmax.relative.size", c(0.01, 0.04)),
   lead.cols    = grDevices::topo.colors(attr(x, "nleaders")),
   vshapes      = c(Leader="square", Follower="circle"),
   add.legend   = TRUE,